Fix typo: rename commponents to components and update all imports
