+++ src/expo-admin/components/operatorSection/OperatorSection.jsx
@@ -10,7 +10,7 @@ import {
   FaTimesCircle,
 } from 'react-icons/fa';
 import styles from './OperatorSection.module.css';
-import ToastSuccess from '../../../common/commponents/toastSuccess/ToastSuccess';
+import ToastSuccess from '../../../common/components/toastSuccess/ToastSuccess';
 
 const mockOperatorData = {
   companyName: '토스트 컴퍼니',
