+++ src/platform-admin/pages/expoCurrent/ExpoCurrent.jsx
@@ -2,9 +2,9 @@ import { useState } from 'react';
 import { FiSearch } from 'react-icons/fi';
 import styles from './ExpoCurrent.module.css';
 
-import Tab from '../../../common/commponents/tab/Tab';
+import Tab from '../../../common/components/tab/Tab';
 import CurrentExpoTable from '../../components/currentExpoTable/CurrentExpoTable';
-import Pagination from '../../../common/commponents/pagination/Pagination';
+import Pagination from '../../../common/components/pagination/Pagination';
 
 function ExpoCurrent() {
   const [currentTab, setCurrentTab] = useState('결제 완료');
